requirements.txt
asgiref==3.2.7
Django==3.0.7
djangorestframework==3.11.0
joblib==0.15.1
numpy==1.18.5
pandas==1.0.4
python-dateutil==2.8.1
pytz==2020.1
scikit-learn==0.23.1
scipy==1.4.1
six==1.15.0
sklearn==0.0
sqlparse==0.3.1
threadpoolctl==2.1.0
tqdm==4.46.1
